Forward and Inverse Linear Sufficient Reduction Methods

Contains Sufficiennt Dimension Reduction methods for Linear dimension reduction. Forward methods available are OPG, OPCG, MAVE and MADE. Inverse methods available include modified versions of SIR, DR and SAVE from Dr. Bing Li's book on SDR (2019). The modifications for Inverse methods allow regularization in a similar spirit to Zhang et. al. (2005).

Details

The DESCRIPTION file: This package was not yet installed at build time.
Index: This package was not yet installed at build time.
~~ An overview of how to use the package, including the most important functions ~~

Author

NA

Maintainer: Harris Quach <harris.quach@gmail.com>

References

~~ Literature or other references for background information ~~

See also

Examples