Set of functions in example RcppArmadillo package

These four functions are created when RcppArmadillo.package.skeleton() is invoked to create a skeleton packages.

rcpparma_hello_world()
rcpparma_outerproduct(x)
rcpparma_innerproduct(x)
rcpparma_bothproducts(x)

Arguments

x

a numeric vector

Value

rcpparma_hello_world() does not return a value, but displays a message to the console. rcpparma_outerproduct() returns a numeric matrix computed as the outer (vector) product of x. rcpparma_innerproduct() returns a double computer as the inner (vector) product of x. rcpparma_bothproducts() returns a list with both the outer and inner products.

Details

These are example functions which should be largely self-explanatory. Their main benefit is to demonstrate how to write a function using the Armadillo C++ classes, and to have to such a function accessible from R.

References

See the documentation for Armadillo, and RcppArmadillo, for more details.

Author

Dirk Eddelbuettel

Examples

  x <- sqrt(1:4)
  rcpparma_innerproduct(x)
#> [1] 10
  rcpparma_outerproduct(x)
#>          [,1]     [,2]     [,3]     [,4]
#> [1,] 1.000000 1.414214 1.732051 2.000000
#> [2,] 1.414214 2.000000 2.449490 2.828427
#> [3,] 1.732051 2.449490 3.000000 3.464102
#> [4,] 2.000000 2.828427 3.464102 4.000000