Sliced Average Variance Estimation

This conducts a sliced average variance estimation (SAVE) as in Li (2018) with modifications to improve speed and to allow for the option of standardizing and regularizing

Usage

save_sdr(x, y, nslices, d, ytype, std = T, lambda = 0)

Arguments

x

a 'n x p' matrix of predictors; n sample size, p dimension

y

a scalar response

nslices

specify the number of slices to conduct;

d

specify the reduced dimension

ytype

specify the response as 'continuous' or 'categorical'

std

should the predictors be standardized? Default is 'TRUE'

lambda

a L2 or Tikonov regularizer for the sample covariance matrix; default is '0', i.e. no regularization

Value

A list containing both the estimate and candidate matrix.

beta - A 'pxd' matrix that estimates a basis for the central subspace.

cand_mat - The candidate matrix for SAVE; this is used in other functions for order determination.

Details

Standardizing is the default as it is necessary for recovering the properly scaled central subspace. However, in certain contexts, the standardization is not necessary, and so we leave this option open to the practitioner.

The L2-regularization option corresponds to the SIR regularization idea by Zhang et al.(2005). While they do not apply the idea to SAVE, we find that the context is analogous and that such a regularization works.